backend/requirements.txt
pywebview>=5.0.5
requests>=2.31.0
